extract file uploader and upload with defined concurreny
